conv c test

+++ tests/conv.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 
 
-void convolute(int output [3][3], int input[5][5], int kernel[3][3] )
+void convolute(double output [3][3], double input[5][5], double kernel[3][3] )
 {
     int convolute = 0; // This holds the convolution results for an index.
     int x, y; // Used for input matrix index
@@ -35,11 +35,11 @@ void convolute(int output [3][3], int input[5][5], int kernel[3][3] )
 void main(int argc, char * argv[]){
 
 
-int input[5][5];
+double input[5][5];
 
 int i =0; int j=0;
 
-int result [3][3]; 
+double result [3][3]; 
 
 int c =0;
 for(i=0;i<5;i++){
@@ -48,7 +48,7 @@ for(i=0;i<5;i++){
     for(j=0;j<5;j++){
 
         input[i][j] = c;
-        printf("%d", input[i][j]);
+        printf("%f", input[i][j]);
         printf("\t");
         c++;
     } 
@@ -56,7 +56,7 @@ for(i=0;i<5;i++){
 
 } 
 
-int kernel [3][3] = {{0,0,0} ,{1,0,-1},{0,0,0}};
+double kernel [3][3] = {{0,0,0} ,{1,0,-1},{0,0,0}};
 
 convolute(result, input, kernel);
 
@@ -66,7 +66,7 @@ for(i=0;i<3;i++){
 
     for(j=0;j<3;j++){
 
-        printf("%d", result[i][j]);
+        printf("%f", result[i][j]);
         
     }
